use_cases/coreset_for_instruction_tuning/requirements.txt
# Coreset Selection for Instruction Tuning
# Core dependencies for the diversity-guided dataset pruning experiment

# Core ML/Data libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
transformers>=4.36.0
datasets>=2.15.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.0

# Embedding generation
sentence-transformers>=2.2.0

# Unsloth for faster training (optional but recommended)
# Install separately: pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
# unsloth

# Evaluation (all local, no paid APIs required)
lm-eval>=0.4.0

# Optional: Local LLM-as-Judge (Prometheus)
# prometheus-eval  # If using Prometheus for evaluation

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Experiment tracking (optional)
wandb>=0.15.0

# For spaCy-based syntactic features (optional)
# spacy>=3.6.0
# python -m spacy download en_core_web_sm
